Fix non-index requests
Because this is a single-page-app built with ember, the application
(currently stored in index.html and assorted javascript files) must be
served whatever the actual URL requested is.

This patch unifies all static assets under /static so they can be served
by a single handler, and creates an extra route for every other
non-matching request to send out the index page.

Fixes #5